Test if qs is ValuesQuerySet before calling values
A nasty effect of .values() in django is that if you call values('field1', 'field2'), you will get a ValuesQuerySet with only those fields. However, if you call values *again* on that qs without args, it will give you values for all fields in the model. This is bad for calling functions that assume by sending a stripped-down values queryset, they can solve the security problem of what fields to export. With this fix, they can as expected.