Force verbose_name to be a unicode object
In some cases, <queryset_instance>.model._meta.fields.verbose_name is
a unicode object as expected, like:
'name' or 'Name of User'

but other in cases, it is a python object representation, like:
<django.utils.functional.__proxy__ object at 0xb2a82bac>

It is unclear why this is happening, but it turned up when trying to
export just the column headers of an empty queryset. By calling unicode,
it is forced to be the desired unicode object.