Fix minor issue in csv assertion
the comparison loop was using zip(), which will only pair values up to the length of the shorter sequence. So in this case, if your test data had 3 rows, but your output data only had 2, the test would still pass even though that's clearly not OK. Using izip_longest() with padding, all elements of the two sequences are paired together.

Also factor out this method so it can be used by another class.