Add support for column headers of related fields
Walking relationships is supported, via qs.values('model__field'), but the field names were not being included in the column header. This is because the name_map was being build from fields in queryset.model._meta.fields (which was done so that verbose_name was accessible), which does not have a concept of relations. This correct first builds from the field names of the queryset, and then overwrites with the verbose names from the model.