Finish adding support for Python 3
Most changes were minor changes to switch to methods that work on both
Python 2 & 3.

The major change was a switch to the underlying CSV library.

Based on a suggestion from the Django documentation I switched to the
python-unicodecsv library, which handles unicode encoding and works
properly with both Python 2 & 3, and is capable of writing to Django's
HttpResponse, which presents an interface of a binary file-like object.
(https://docs.djangoproject.com/en/1.9/howto/outputting-csv/#using-the-python-csv-library)