Update version matrix in .travis.yml and update test tools
The main motivation for this change was to add testing for Django 3.0. While
adding that version we also add the interim Django versions and Python 3.x
versions. Several of these are not supported, but testing on these older
versions may allow an application that is stuck on an old version for financial
or other reasons to still consider adding djqscsv.

Python 3.5 is no longer installable so it is excluded.

Supporting Python 3.8 required updated versions of flake8 and coverage. We opted
to upgrade to the latest versions because they still have support back to Python
2.7.